requirements.txt
google-auth-oauthlib
google-api-python-client
google-generativeai
SpeechRecognition
edge-tts
pygame
beautifulsoup4
python-dotenv